Allow migrations to apply correctly to a fresh database.
In non-legacy (fresh) databases, the djcelery tables have never been
created, so dropping them unconditinally fails.

+++ ecs/core/migrations/0009_remove_djcelery.py
@@ -9,7 +9,7 @@ class Migration(migrations.Migration):
 
     operations = [
         migrations.RunSQL('''
-            drop table
+            drop table if exists
                 djcelery_crontabschedule,
                 djcelery_intervalschedule,
                 djcelery_periodictask,
